MCPLangchain/MCPDemo - CustomeServer/pyproject.toml
[project]
name = "mcpdemo-customeserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastmcp>=2.14.1",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
]
